refacor fetchAllCommit function
