+++ src/handlePRUpdated.ts
@@ -1,67 +1,5 @@
 import * as github from '@actions/github';
-import * as core from '@actions/core';
-
-interface Commit {
-  sha: string;
-  commit: {
-    message: string;
-    author: {
-      name: string;
-    };
-  };
-  author: {
-    login: string;
-  } | null;
-}
-
-async function fetchAllCommits(
-  owner: string,
-  repo: string,
-  pullNumber: number,
-  githubToken: string
-): Promise<Commit[]> {
-  const allCommits: Commit[] = [];
-  let url:
-    | string
-    | null = `https://api.github.com/repos/${owner}/${repo}/pulls/${pullNumber}/commits?per_page=100`;
-  let page = 1;
-
-  while (url) {
-    core.info(`Fetching page ${page}: ${url}`);
-    const response: Response = await fetch(url, {
-      headers: {
-        Authorization: `token ${githubToken}`,
-      },
-    });
-
-    if (!response.ok) {
-      const errorData = await response.json();
-      throw new Error(
-        `GitHub API request failed: ${response.status} ${
-          response.statusText
-        } - ${JSON.stringify(errorData)}`
-      );
-    }
-
-    const commitsData: Commit[] = await response.json();
-    core.info(`Fetched ${commitsData.length} commits on page ${page}`);
-    allCommits.push(...commitsData);
-
-    const linkHeader: string | null = response.headers.get('link');
-    core.info(`Link Header: ${linkHeader}`);
-    if (linkHeader) {
-      const nextLinkMatch = linkHeader.match(/<([^>]+)>;\s*rel="next"/);
-      url = nextLinkMatch ? nextLinkMatch[1] : null;
-    } else {
-      url = null;
-    }
-
-    page++;
-  }
-
-  core.info(`Fetched a total of ${allCommits.length} commits`);
-  return allCommits;
-}
+import { fetchAllCommits, Commit } from './utils/fetchAllCommits';
 
 export async function handlePRUpdated(
   slackToken: string,
